Renamed test project for authentication and tidy it up a bit.
