+++ src/Altinn.Platform/Altinn.Platform.Authentication/Altinn.Platform.Authentication.Tests/Altinn.Platform.Authentication.Tests.csproj
@@ -16,7 +16,6 @@
     <DocumentationFile>bin\Debug\netcoreapp3.1\Altinn.Platform.Authentication.UnitTest.xml</DocumentationFile>
   </PropertyGroup>
 
-
   <ItemGroup>
     <Content Include="appsettings.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
@@ -32,9 +31,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.1.4" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.1.3" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.6.1" />
-    <PackageReference Include="Moq" Version="4.14.1" />
+    <PackageReference Include="Moq" Version="4.14.0" />
     <PackageReference Include="StyleCop.Analyzers" Version="1.1.118">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
@@ -52,10 +51,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Authentication\Altinn.Platform.Authentication.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Fixtures\" />
   </ItemGroup>
 
   <ItemGroup>
