+++ src/Altinn.Platform/Altinn.Platform.Authentication/Altinn.Platform.Authentication.Tests/Fakes/ConfigurationManagerStub.cs
@@ -7,26 +7,13 @@ using Microsoft.IdentityModel.Protocols;
 using Microsoft.IdentityModel.Protocols.OpenIdConnect;
 using Microsoft.IdentityModel.Tokens;
 
-namespace Altinn.Platform.Authentication.IntegrationTests.Fakes
+namespace Altinn.Platform.Authentication.Tests.Fakes
 {
     /// <summary>
     /// Represents a stub of <see cref="ConfigurationManager{OpenIdConnectConfiguration}"/> to be used in integration tests.
     /// </summary>
     public class ConfigurationManagerStub : IConfigurationManager<OpenIdConnectConfiguration>
     {
-        /// <summary>
-        /// Initializes a new instance of <see cref="ConfigurationManagerStub" />
-        /// </summary>
-        /// <param name="metadataAddress">The address to obtain configuration.</param>
-        /// <param name="configRetriever">The <see cref="IConfigurationRetriever{OpenIdConnectConfiguration}" /></param>
-        /// <param name="docRetriever">The <see cref="IDocumentRetriever" /> that reaches out to obtain the configuration.</param>
-        public ConfigurationManagerStub(
-            string metadataAddress,
-            IConfigurationRetriever<OpenIdConnectConfiguration> configRetriever,
-            IDocumentRetriever docRetriever)
-        {
-        }
-
         /// <inheritdoc />
         public async Task<OpenIdConnectConfiguration> GetConfigurationAsync(CancellationToken cancel)
         {
