+++ src/Altinn.Platform/Altinn.Platform.Authentication/Altinn.Platform.Authentication.Tests/Repositories/OrgLookupTest.cs
@@ -6,10 +6,11 @@ using Altinn.Platform.Authentication.Repositories;
 
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+
 using Moq;
 using Xunit;
 
-namespace Altinn.Platform.Authentication.IntegrationTests
+namespace Altinn.Platform.Authentication.Tests.Repositories
 {
     /// <summary>
     /// Testclass for harvesting orgs.
@@ -22,15 +23,12 @@ namespace Altinn.Platform.Authentication.IntegrationTests
         [Fact]
         public async Task TestHarvestOrgs_OK()
         {
-            Mock<IOptions<GeneralSettings>> optionsMock = new Mock<IOptions<GeneralSettings>>();
             Mock<ILogger<OrganisationRepository>> loggerMock = new Mock<ILogger<OrganisationRepository>>();
-            Mock<GeneralSettings> generalSettingsMock = new Mock<GeneralSettings>();
-            GeneralSettings generalSettings = new GeneralSettings();
-            generalSettings.OrganisationRepositoryLocation = "https://altinncdn.no/orgs/altinn-orgs.json";
 
-            optionsMock
-                .Setup(o => o.Value)
-                .Returns(generalSettings);
+            GeneralSettings generalSettings = new GeneralSettings { OrganisationRepositoryLocation = "https://altinncdn.no/orgs/altinn-orgs.json" };
+
+            Mock<IOptions<GeneralSettings>> optionsMock = new Mock<IOptions<GeneralSettings>>();
+            optionsMock.Setup(o => o.Value).Returns(generalSettings);
 
             OrganisationRepository orgRepo = new OrganisationRepository(loggerMock.Object, optionsMock.Object);
