+++ src/Altinn.Platform/Altinn.Platform.Authentication/Altinn.Platform.Authentication.Tests/selfSignedEncryptionTestCertificate.pfx
