+++ src/Altinn.Platform/Altinn.Platform.Authentication/Altinn.Platform.Authentication.Tests/selfSignedEncryptionTestCertificatePublic.cer
