+++ src/Altinn.Platform/Altinn.Platform.Authentication/Altinn.Platform.Authentication.sln
@@ -12,7 +12,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Altinn.Platform.Authentication", "Authentication\Altinn.Platform.Authentication.csproj", "{A7470874-7D29-4526-A9C6-4133595C1182}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Altinn.Platform.Authentication.IntegrationTests", "Altinn.Platform.Authentication.IntegrationTests\Altinn.Platform.Authentication.IntegrationTests.csproj", "{C30FD85C-F773-4BF8-A65C-14CCF977A76A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Altinn.Platform.Authentication.Tests", "Altinn.Platform.Authentication.Tests\Altinn.Platform.Authentication.Tests.csproj", "{BBEE9804-91DA-4244-9795-A49D83465D52}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -28,10 +28,10 @@ Global
 		{A7470874-7D29-4526-A9C6-4133595C1182}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A7470874-7D29-4526-A9C6-4133595C1182}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A7470874-7D29-4526-A9C6-4133595C1182}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C30FD85C-F773-4BF8-A65C-14CCF977A76A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C30FD85C-F773-4BF8-A65C-14CCF977A76A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C30FD85C-F773-4BF8-A65C-14CCF977A76A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C30FD85C-F773-4BF8-A65C-14CCF977A76A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BBEE9804-91DA-4244-9795-A49D83465D52}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BBEE9804-91DA-4244-9795-A49D83465D52}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BBEE9804-91DA-4244-9795-A49D83465D52}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BBEE9804-91DA-4244-9795-A49D83465D52}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
